Basic Database Access Finished
The basic database access functions are finished in theory but not yet
tested. These functions are yet to be implemented in Set and Card since
no way to add sets yet exists. A fragment for creating a set was made,
and another fragment for editing a set needs to be made as well.